deepseek-v3-implementation/requirements.txt
# Core dependencies
torch>=2.1.0
numpy>=1.24.0
transformers>=4.35.0

# Distributed training
deepspeed>=0.12.0
megatron-core>=0.4.0

# DeepSeek components (install from source)
# FlashMLA: https://github.com/deepseek-ai/FlashMLA
# DeepEP: https://github.com/deepseek-ai/DeepEP

# Flash Attention (required for FlashMLA)
flash-attn>=2.3.0

# Data processing
datasets>=2.14.0
tokenizers>=0.15.0
pyarrow>=14.0.0          # For Parquet format support
ftfy>=6.1.0              # Fix text encoding issues (preliminary cleaning)
datasketch>=1.6.0        # MinHash LSH for deduplication (Lee et al. 2022)
langdetect>=1.0.9        # Language detection
fasttext-wheel>=0.9.2    # Quality classification (Xu et al. 2024)
regex>=2023.10.0         # Advanced regex for text processing
zstandard>=0.21.0        # Compression support for datasets

# Monitoring and logging
wandb>=0.16.0
tensorboard>=2.15.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
packaging>=23.0

# Optimization
ninja>=1.11.0  # For faster compilation
triton>=2.1.0  # For custom kernels

# Optional: Debugging and profiling
py-spy>=0.3.14
nvtx>=0.2.5
